refactor(hf-154): dedupe clipboard helper + strip paired block-component tags
- Extract the shared Clipboard-API-with-textarea-fallback logic from
  CopyMarkdownButton.vue and CodingAgentWizard.vue into a single
  components/clipboard.js (copyToClipboard); both now import it.
- strip.js: also drop a whole-line closing tag of a paired block
  component (</Foo>); the opening tag and inner content were already
  handled, only the stray closing tag leaked. +1 spec assertion (13).
