+++ docs/.vuepress/components/CodingAgentWizard.vue
@@ -22,6 +22,8 @@
 </template>
 
 <script>
+import { copyToClipboard } from './clipboard';
+
 export default {
   name: 'CodingAgentWizard',
   data() {
@@ -65,23 +67,10 @@ export default {
     reset() { this.selected = null; this.copied = false; },
     copy() {
       if (!this.current) return;
-      const text = this.current.snippet;
-      const fallback = (t) => {
-        const el = document.createElement('textarea');
-        el.value = t;
-        el.style.position = 'fixed';
-        el.style.opacity = '0';
-        document.body.appendChild(el);
-        el.select();
-        try { document.execCommand('copy'); } finally { document.body.removeChild(el); }
-      };
-      const done = () => { this.copied = true; setTimeout(() => { this.copied = false; }, 1500); };
-      if (navigator.clipboard && window.isSecureContext) {
-        navigator.clipboard.writeText(text).then(done).catch(() => { fallback(text); done(); });
-      } else {
-        fallback(text);
-        done();
-      }
+      copyToClipboard(this.current.snippet).then(() => {
+        this.copied = true;
+        setTimeout(() => { this.copied = false; }, 1500);
+      });
     },
   },
 };
