+++ docs/.vuepress/components/CopyMarkdownButton.vue
@@ -9,6 +9,8 @@
 </template>
 
 <script>
+import { copyToClipboard } from './clipboard';
+
 export default {
   name: 'CopyMarkdownButton',
   data() {
@@ -27,22 +29,10 @@ export default {
   methods: {
     copy() {
       const absolute = window.location.origin + this.mdUrl;
-      const fallback = (text) => {
-        const el = document.createElement('textarea');
-        el.value = text;
-        el.style.position = 'fixed';
-        el.style.opacity = '0';
-        document.body.appendChild(el);
-        el.select();
-        try { document.execCommand('copy'); } finally { document.body.removeChild(el); }
-      };
-      const done = () => { this.copied = true; setTimeout(() => { this.copied = false; }, 1500); };
-      if (navigator.clipboard && window.isSecureContext) {
-        navigator.clipboard.writeText(absolute).then(done).catch(() => { fallback(absolute); done(); });
-      } else {
-        fallback(absolute);
-        done();
-      }
+      copyToClipboard(absolute).then(() => {
+        this.copied = true;
+        setTimeout(() => { this.copied = false; }, 1500);
+      });
     },
   },
 };
